% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gql.R
\name{ess_catalogue_refresh}
\alias{ess_catalogue_refresh}
\title{Forget cached catalogue answers}
\usage{
ess_catalogue_refresh()
}
\value{
Invisibly, the number of cached answers discarded.
}
\description{
Catalogue lookups are remembered for the duration of the session. Call this
to discard them, for instance after a new ESS round has been published while
your session was open.
}
\details{
This does not touch downloaded data files; for those see \code{\link[=ess_cache_clear]{ess_cache_clear()}}.
}
\examples{
ess_catalogue_refresh()

}
